Conflicts:
	plugins/de.ovgu.featureide.fm.core/src/de/ovgu/featureide/fm/core/io/velvet/VelvetLexer.java (replaced with ours)
	plugins/de.ovgu.featureide.fm.core/src/de/ovgu/featureide/fm/core/io/velvet/VelvetParser.java (replaced with ours)
	plugins/de.ovgu.featureide.fm.ui/src/de/ovgu/featureide/fm/ui/editors/FeatureModelEditor.java (change added)

+++ plugins/de.ovgu.featureide.ui/src/de/ovgu/featureide/ui/statistics/core/composite/lazyimplementations/DirectivesNode.java
@@ -115,7 +115,7 @@ public class DirectivesNode extends LazyParent {
 					average += i;
 				}
 				if (list.size() != 0) {
-					average /= (double) list.size();
+					average /= list.size();
 					average *= 10;
 					long rounded = Math.round(average);
 					average = ((double) rounded) / 10;
@@ -164,7 +164,7 @@ public class DirectivesNode extends LazyParent {
 			
 			Integer numberOfChildren = parent.getChildren().length;
 			
-			double average = (double) numberOfDirectives;
+			double average = numberOfDirectives;
 			
 			average /= (double) numberOfChildren;
 			average *= 10;
